requirements.txt
action==1.4.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
Django==2.2.2
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.7.7
drf-yasg==1.16.0
idna==2.8
inflection==0.3.1
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
requests==2.22.0
ruamel.yaml==0.15.97
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
